Avoid "warning: missing initializer for member” [-Wmissing-field-initializers]
Based on rationale given in https://stackoverflow.com/questions/21658485/why-do-i-get-warning-missing-initializer-for-member-wmissing-field-initial.

Using {} for initializing a struct is not good enough since we are not given any value to it, and hence the gcc warning. Doing {0} just sets the value of the first field in the struct, so it's not good either.