Channel: move some ids from internal object to data object
For instance, when calling `transport.consume(options)`, the user must pass `producerId` within the given options. When it comes to route the channel request to the corresponding `Transport`, such a `producerId` is not used for "internal routing" and hence it must be in the `dta` object instead.
In fact, we were already doing it properly in `rtpObserver.addProducer(options)` in which `options.producerId` was sent to the worker within the `request.data` object rather than `request.internal`.

Changes in Node layer:

- `router.createWebRtcTransportWithServer()`: Move `webRtcServerId` from `internal` to `data`.
- `transport.consume()`: Move `producerId` from `internal` to `data`. Also in `PipeTransport.consume()` method.
- `transport.consumeData()`: Move `dataProducerId` from `internal` to `data`.

Changes in Rust layer:

- In theory similar changes.
- **TODO:** Change related to `webRtcServerId`.
- **TODO:** The "smoke" test fails so probably I've broken something :(
   > thread 'smoke::smoke' panicked at 'called `Result::unwrap()` on an `Err` value:
   >   Response { reason: "missing internal.producerId [method:producer.enableTraceEvent]" }',
   >   rust/tests/integration/smoke.rs:203:18