SimulcastConsumer::GetDesiredBitrate(): Choose the highest bitrate among all Producer streams
Instead of assuming that the highest simulcast stream is the one consuming highest bitrate, iterate all them and pick up the highest value.

This is because wWhen the sender enables a higher stream, initially the bitrate of it could be less than the bitrate of a lower stream. In a few seconds it will become higher. However don't report its irrelevant current bitrate in the meantime.