SeqManager: Properly remove dropped entries (#1054)
* SeqManager: Properly remove dropped entries

Fixes #1037.

Keep dropped entries clean by removing those higher than this->maxInput.

* Set dropped input to 'maxInput' if needed

A dropped input is part of the same sequence, hence set it to 'maxInput'
if the dropped value is higher than the current 'maxInput'.

Without this consideration we would be wrongly erasing dropped inputs
that are legitimally just the maximum input of the sequence stream.

* Update CHANGELOG

* Input(): Performance, do not loop over every dropped inputs

Which has a linear complexity. Use lower_bound which has logarithmic
complexity.