Only perform RTP inactivity checks on simulcast streams
We compute the received stream `score` based on many factors, including whether packets are being redeived or not. If the encoder in client side stops sending packets we (in mediasoup) assume that something is wrong (RTP inactivity timer fires) and set score to 0. This PR changes this behavior and only enables the RTP inactivity check on simulcast streams. Why? Because in simulcast we NEED this in order to detect when the higher streams stop so we can quickly switch consumers to the lower stream.