Refactor data consumer SCTP stream parameter handling
The API of `DataConsumer` and `DataProducer` have been unified to use the same
kind of parameters object when providing SCTP stream parameters and those
parameters are always validated.

Previously, creating a `DataConsumer` from a `DataProducer` that has been
created via `DirectTransport` was brittle since SCTP stream parameters cannot
be inferred from the `DirectTransport`'s `DataProducer`.

It is now possible to provide the SCTP stream id for the consumer, too.