Simplify meson setup task
### Details

- Remove try/except logic when running `meson setup`.
- Do not use `--reconfigure`.

### Rationale

Previous code was a workaround for a potential issue https://github.com/ninja-build/ninja/issues/1997 in ninja, however I cannot reproduce it. This is what I've tried:

1. I've removed the `--reconfigure` in the `meson setup` task.
2. I've run `invoke` which runs `mediasoup-worker` task which runs `setup` task. Everything is built properly.
3. I've added a new file `worker/src/foo.cpp` with some random content and added it in `meson.build`.
4. I've run `invoke` again which, according to the issue reported in ninja project, should now fail due to "ModuleNotFoundError: No module named 'mesonbuild'". However it doesn't fail.

Perhaps the issue was auto-magically fixed in some recent version of meson or ninja and we didn't notice it when we upgrade those in mediasoup in the recent past.

BTW, even if the issue still exists when manually editing `meson,build` (I couldn't reproduce it as told above), the user could just call `invoke` with `MESON_ARGS="--reconfigure" invoke`. We don't need all those try/except with 99% duplicate code.