rust: Implement feature flag to disable worker build
The `mediasoup` crate may need to be included to allow for using prelude
or other defined types without the need to include the `mediasoup-sys`
bindings to `libmediasoup-worker`.

Place the compilation of `mediasoup-sys` behind a feature flag that is
enabled by defaault. When the feature is disabled, build speeds are
dramatically increased and the build dependencies of the library are not
needed in the environment.