Worker: modernize clang-format
We were using a very old clang-format out of npm [clang-format package](https://www.npmjs.com/package/clang-format)
which is out of date.

Use a modern clang-format and define a specific version to be run.

Also just run lint GH action for macos. We are not testing clang-format on different OS.