Worker: Perf, avoid dynamic memory allocation in TransportTuple hash
For every packet arriving from network we are generating a tuple, and upon creation
we create its hash.

Before this PR we were using a std::vector which we later fill it with the tuple info.
This generates dynamic memory allocation which must later be released.

This PR makes use of std::array with a fixed length (we know it in advance).

This only changes improves CPU performance in 1%.