use the old good array
std::array does not provide the number of added elements, just the total size.

We need it in order to call GenerateFnv1aHash() with the proper size. Previous to this
commit we were passing the whole array capacity rather than the real size, meaning
extra length for IPv4 tuples (it does not hurt, but it's not correct either).

Since we need an incremental index anyway there is no reason to keep using std::array.